refactor(code-edit): the declaration editor is its own package
The AST-locate-then-text-splice editor behind console:updateFunctionConfig and
console:updateAgentConfig moves out of @pikku/addon-console into
@pikku/code-edit. Pure move: same class, same behaviour, same scopes, five
import repoints and one dependency.

Editing a pikku declaration is not a console concern. The CLI needs the same
operation, and a private service inside an addon leaves copying as the only way
to get it. It stays out of @pikku/inspector on purpose — the inspector is the
read model codegen and validation depend on precisely because it does not
mutate, so a file writer there would make every read-path consumer link one.

The addon still resolves it through the lazy variable-path dynamic import, so
self-contained bundles that never edit code still do not ship the TypeScript
compiler. The service only ever calls ts.createSourceFile — a parse of the one
file being edited, no program and no type checker.

Two things the move fixes for free: the 35 tests were sitting in
@pikku/addon-console, which the root test script excludes, so they never ran in
CI and now do; and the package is MIT, matching the addon it came from.
