+++ packages/addon/pikku-console/package.json
@@ -52,6 +52,7 @@
     "typescript": "^6.0.3"
   },
   "dependencies": {
+    "@pikku/code-edit": "^0.12.0",
     "@pikku/knowledge": "^0.12.7",
     "json-schema": "^0.4.0",
     "open": "^11.0.1"
