+++ packages/addon/pikku-console/src/functions/update-function-config.function.ts
@@ -1,6 +1,6 @@
 import { LocalEnvironmentOnlyError } from '#pikku/addon/error'
 import { pikkuFunc } from '#pikku/addon/function'
-import type { FunctionConfigChanges } from '../services/code-edit.service.js'
+import type { FunctionConfigChanges } from '@pikku/code-edit'
 
 export const updateFunctionConfig = pikkuFunc<
   {
