+++ packages/addon/pikku-console/types/application-types.d.ts
@@ -8,7 +8,7 @@ import type { MetaService } from '@pikku/core/services'
 import type { WiringService } from '../src/services/wiring.service.js'
 import type { AddonService } from '../src/services/addon.service.js'
 import type { AddonReadinessService } from '../src/services/addon-readiness.service.js'
-import type { CodeEditService } from '../src/services/code-edit.service.js'
+import type { CodeEditService } from '@pikku/code-edit'
 import type { StateDiffService } from '../src/services/state-diff.service.js'
 import type { DbSchemaService } from '../src/services/db-schema.service.js'
 import type { KnowledgeService } from '../src/services/knowledge.service.js'
