+++ packages/code-edit/src/index.ts
@@ -1,0 +1,1 @@
+export * from './code-edit.service.js'
